refactor: fix folders names
